Move cacheFetch to runtime
