+++ packages/runtime/src/index.ts
@@ -22,5 +22,16 @@ export { startPlayground, stopPlayground, defaultWpCliPath } from "./playground.
 // Export WP CLI utilities
 export { wpCli, getWordPressVersion } from "./wp-cli.js";
 
+// Export caching utilities
+export { cacheFetch } from "./cache-fetch.js";
+export type { CacheFetchOptions } from "./cache-fetch.js";
+export {
+  CACHE_FOREVER,
+  CACHE_DISABLED,
+  CACHE_1_HOUR,
+  CACHE_1_DAY,
+  CACHE_1_WEEK,
+} from "./cache-fetch.js";
+
 // Export HTTP utilities
 export { request } from "./request.js";
